+++ presentation/src/main/java/com/example/presentation/ui/gallery/GalleryViewModel.kt
@@ -3,7 +3,6 @@ package com.example.presentation.ui.gallery
 import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
-import com.example.presentation.ui.gallery.GalleryStore.*
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
