--- presentation/src/main/java/mvi/Reducer.kt
@@ -1,6 +1,0 @@
-package mvi
-
-fun interface Reducer<STATE, MESSAGE> {
-
-    fun reduce(state: STATE, message: MESSAGE): STATE
-}
